Refactor: - 로또 테스트 네이밍 바뀌는 문제등 해결
